nbt, faster chunk rendering, jack block
that and improved chunk searcher along with view distance option

+++ src/net/oikmo/network/shared/PacketChunk.java
@@ -3,7 +3,7 @@ package net.oikmo.network.shared;
 public class PacketChunk {
 
 	public int id;
-	public float x;
-	public float z;
+	public int x;
+	public int z;
 	public byte[] data;
 }
